Healthy Programmer
This is a program that runs in background and reminds the user to drink water , do eye exercise and physical exercise at stipulated intervals through alarms and also maintains log for the same. It can be modified as a reminder for many other things like ranging from medicines to period dates.